third_party/blink/web_tests/external/wpt/svg/linking/scripted/href-mpath-element-expected.txt

This is a testharness.js-based test.
[FAIL] Test for mpath when removing href but we still have xlink:href
  assert_equals: expected 0 but got 50
Harness: the test ran to completion.
